\name{imis}
\alias{imis}
\docType{data}
\title{
Data from an experiment using incomplete split-plot design
}
\description{
Data from an experiment using incomplete split-plot design where blocks are incomplete with respect to main plot treatments and main plots are also incomplete with respect to subplot treatments
}
\usage{data("imis")}
\format{
  A data frame with 36 observations on the following 5 variables.
  \describe{
    \item{\code{obs}}{
Observations}
    \item{\code{block}}{Blocks}
	 \item{\code{mp}}{Main plot treatments}
	  \item{\code{sp}}{Subplot treatments}
    \item{\code{y}}{The response variable}
  }
}

\examples{
data(imis)
}
\keyword{datasets}
