\name{ispd.cmis}
\alias{ispd.cmis}
\title{Incomplete split-plot design for given number of main plot treatments, number of subplot treatments, number of blocks and number of subplot treatments in each main plot}
\description{This function generate an incomplete split-plot design for given number of main plot treatments, number of subplot treatments, number of blocks and number of subplot treatments in each main plot}
\usage{ispd.cmis(v1,v2,b,k2)}
\arguments{
	\item{v1}{number of main plot treatments}
	\item{v2}{number of subplot treatments}
	\item{b}{number of block}	
	\item{k2}{number of subplot treatments in each main plot}
	}
\value{A list containing parameters, design layout and column layout of design}
\author{Baidya Nath Mandal <mandal.stat@gmail.com>}
\keyword{internal}