\name{ispd.imcs}
\alias{ispd.imcs}
\title{Incomplete split-plot design for given number of main plot treatments, number of subplot treatments, number of blocks and number of main plot treatments in each block}
\description{This function generate an incomplete split-plot design for given number of main plot treatments, number of subplot treatments, number of blocks and number of main plot treatments in each block}
\usage{ispd.imcs(v1,b,k1,v2)}
\arguments{
	\item{v1}{number of main plot treatments}
	\item{b}{number of blocks}	
	\item{k1}{number of main plot treatments in each block}
	\item{v2}{number of subplot treatments}	
	}
\value{A list containing parameters, design layout and column layout of design}
\author{Baidya Nath Mandal <mandal.stat@gmail.com>}
\keyword{internal}