\name{spd}
\alias{spd}
\title{Complete split-plot design for given number of main plot treatments, number of subplot treatments and number of blocks}
\description{This function generate a complete split-plot design for given number of main plot treatments, number of subplot treatments and number of blocks}
\usage{spd(v1,v2,b)}
\arguments{
	\item{v1}{number of main plot treatments}
	\item{v2}{number of subplot treatments}
	\item{b}{number of block}
    }
\value{A list containing parameters, design layout and column layout of design}
\author{Baidya Nath Mandal <mandal.stat@gmail.com>}
\keyword{internal}